splitting up some code into other modules, cleaned up so that the IDE reports 0 errors
