+++ project/queries/constants.py
@@ -1,0 +1,19 @@
+GRANTS = 'grants'
+KEYWORDS = 'keywords'
+PAPERS = 'papers'
+PEOPLE = 'people'
+ORGANIZATIONS = 'organizations'
+DATES = 'dates'
+GRANT_DATE_RANGE = 'grantDateRange'
+PUBLICATION_DATE_RANGE = 'publicationDateRange'
+GRANT_LIST = 'grantList'
+PAPER_LIST = 'paperList'
+START = 'start'
+END = 'end'
+
+withPeople = 'withPeople'
+withOrgs = 'withOrgs'
+withKeywords = 'withKeywords'
+withPapers = 'withPapers'
+withDates = 'withDates'
+withGrants = 'withGrants'
